Fixed some warnings.
Can't use uint64_t as a bitfield; let's use unsigned int instead.

+++ x86/gdt.c
@@ -12,18 +12,18 @@
 */
 typedef struct gdt_t gdt_t;
 struct gdt_t {
-	uint64_t limit_lo  : 16; 
-	uint64_t base_lo   : 24; 
-	uint64_t type      :  4; /* segment type */
-	uint64_t s         :  1; /* descriptor type */
-	uint64_t dpl       :  2; /* descriptor priviledge level */
-	uint64_t p         :  1; /* present? */
-	uint64_t limit_hi  :  4;
-	uint64_t avl       :  1; /* available */
-	uint64_t l         :  1; /* 64-bit code segment (i.e. Long mode) */
-	uint64_t db        :  1; /* default operation size (16 vs 32 bit) */
-	uint64_t g         :  1; /* granularity */
-	uint64_t base_hi   :  8;
+	unsigned int limit_lo  : 16; 
+	unsigned int base_lo   : 24; 
+	unsigned int type      :  4; /* segment type */
+	unsigned int s         :  1; /* descriptor type */
+	unsigned int dpl       :  2; /* descriptor priviledge level */
+	unsigned int p         :  1; /* present? */
+	unsigned int limit_hi  :  4;
+	unsigned int avl       :  1; /* available */
+	unsigned int l         :  1; /* 64-bit code segment (i.e. Long mode) */
+	unsigned int db        :  1; /* default operation size (16 vs 32 bit) */
+	unsigned int g         :  1; /* granularity */
+	unsigned int base_hi   :  8;
 	
 }__attribute__((packed));
 _Static_assert(sizeof(gdt_t) == 8, "GDT packing error");
